@@ -66,12 +66,12 @@
     "pg": "^4.5.6",
     "pouchdb": "^7.0.0",
     "prop-types": "^15.5",
-    "react": "^15.5",
+    "react": "^16.0",
     "react-addons-css-transition-group": "^15.5",
     "react-addons-pure-render-mixin": "^15.5",
     "react-c3js": "^0.1.19",
     "react-datetime": "^2.11.0",
-    "react-dom": "^15.5",
+    "react-dom": "^16.0",
     "reactify": "^1.1.1",
     "redis": "^2.6.2",
     "request": "^2.73.0",
@@ -81,7 +81,8 @@
     "tablesorter": "^2.29.0",
     "underscore": "^1.8.3",
     "winston": "^2.3.1",
-    "wurl": "^2.5.2"
+    "wurl": "^2.5.2",
+    "yarn": "^1.22.19"
   },
   "devDependencies": {
     "assert": "~1.3.0",
@@ -104,7 +105,7 @@
     "eslint-plugin-import": "^2.2.0",
     "fancy-log": "^1.3.2",
     "fs-extra": "^7.0.1",
-    "gulp": "^3.9.1",
+    "gulp": "3.9.1",
     "gulp-babel": "^7.0.0",
     "gulp-less": "^3.3.2",
     "gulp-sourcemaps": "^2.6.1",
