@@ -7,14 +7,16 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import 'leaflet/dist/leaflet.css';
 import 'jquery-ui/themes/base/all.css';
 
+
 function loadPage(){
     var path = url(1) == 'portal' ? url(2) : url(1);
+    
     switch(path){
         case '':
             require('./home');
             break;
         case 'search':
-            require('./search');
+            require('./search')
             break;
         case 'tutorial':
             require('./tutorial');
